login/welcome/serializers.py
from .models import Users
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Users
        fields = ['firstname', 'lastname', 'email']
